MNT refactor the project structure (#27)
* MNT refactor the project structure
